+++ ai/instill/v0/image_classification.go
@@ -10,9 +10,9 @@ import (
 	modelPB "github.com/instill-ai/protogen-go/model/model/v1alpha"
 )
 
-func (e *execution) executeImageClassification(grpcClient modelPB.ModelPublicServiceClient, modelName string, inputs []*structpb.Struct) ([]*structpb.Struct, error) {
+func (e *execution) executeImageClassification(grpcClient modelPB.ModelPublicServiceClient, nsID string, modelID string, version string, inputs []*structpb.Struct) ([]*structpb.Struct, error) {
 	if len(inputs) <= 0 {
-		return nil, fmt.Errorf("invalid input: %v for model: %s", inputs, modelName)
+		return nil, fmt.Errorf("invalid input: %v for model: %s/%s/%s", inputs, nsID, modelID, version)
 	}
 
 	if grpcClient == nil {
@@ -33,18 +33,18 @@ func (e *execution) executeImageClassification(grpcClient modelPB.ModelPublicSer
 		taskInputs = append(taskInputs, &modelPB.TaskInput{Input: taskInput})
 	}
 
-	taskOutputs, err := trigger(grpcClient, e.SystemVariables, modelName, taskInputs)
+	taskOutputs, err := trigger(grpcClient, e.SystemVariables, nsID, modelID, version, taskInputs)
 	if err != nil {
 		return nil, err
 	}
 	if len(taskOutputs) <= 0 {
-		return nil, fmt.Errorf("invalid output: %v for model: %s", taskOutputs, modelName)
+		return nil, fmt.Errorf("invalid output: %v for model: %s/%s/%s", taskOutputs, nsID, modelID, version)
 	}
 	outputs := []*structpb.Struct{}
 	for idx := range inputs {
 		imgClassificationOp := taskOutputs[idx].GetClassification()
 		if imgClassificationOp == nil {
-			return nil, fmt.Errorf("invalid output: %v for model: %s", imgClassificationOp, modelName)
+			return nil, fmt.Errorf("invalid output: %v for model: %s/%s/%s", imgClassificationOp, nsID, modelID, version)
 		}
 		outputJSON, err := protojson.MarshalOptions{
 			UseProtoNames:   true,