+++ application/website/v0/main.go
@@ -49,12 +49,12 @@ func Init(bc base.Component) *component {
 	return comp
 }
 
-func (c *component) CreateExecution(sysVars map[string]any, setup *structpb.Struct, task string) (*base.ExecutionWrapper, error) {
+func (c *component) CreateExecution(x base.ComponentExecution) (base.IExecution, error) {
 	e := &execution{
-		ComponentExecution: base.ComponentExecution{Component: c, SystemVariables: sysVars, Setup: setup, Task: task},
+		ComponentExecution: x,
 	}
 
-	switch task {
+	switch x.Task {
 	case taskScrapeWebsite:
 		e.execute = e.Scrape
 	case taskScrapeSitemap:
@@ -62,10 +62,10 @@ func (c *component) CreateExecution(sysVars map[string]any, setup *structpb.Stru
 		e.externalCaller = scrapSitemapCaller
 		e.execute = e.ScrapeSitemap
 	default:
-		return nil, fmt.Errorf(task + " task is not supported.")
+		return nil, fmt.Errorf(x.Task + " task is not supported.")
 	}
 
-	return &base.ExecutionWrapper{Execution: e}, nil
+	return e, nil
 }
 
 func (e *execution) Execute(_ context.Context, inputs []*structpb.Struct) ([]*structpb.Struct, error) {