+++ data/pinecone/v0/component_test.go
@@ -192,7 +192,7 @@ func TestComponent_Execute(t *testing.T) {
 	}
 
 	bc := base.Component{}
-	connector := Init(bc)
+	cmp := Init(bc)
 
 	for _, tc := range testcases {
 		c.Run(tc.name, func(c *qt.C) {
@@ -225,13 +225,17 @@ func TestComponent_Execute(t *testing.T) {
 				"url":     pineconeServer.URL,
 			})
 
-			exec, err := connector.CreateExecution(nil, setup, tc.task)
+			exec, err := cmp.CreateExecution(base.ComponentExecution{
+				Component: cmp,
+				Setup:     setup,
+				Task:      tc.task,
+			})
 			c.Assert(err, qt.IsNil)
 
 			pbIn, err := base.ConvertToStructpb(tc.execIn)
 			c.Assert(err, qt.IsNil)
 
-			got, err := exec.Execution.Execute(ctx, []*structpb.Struct{pbIn})
+			got, err := exec.Execute(ctx, []*structpb.Struct{pbIn})
 			c.Check(err, qt.IsNil)
 
 			c.Assert(got, qt.HasLen, 1)
@@ -255,11 +259,15 @@ func TestComponent_Execute(t *testing.T) {
 			"url": pineconeServer.URL,
 		})
 
-		exec, err := connector.CreateExecution(nil, setup, taskUpsert)
+		exec, err := cmp.CreateExecution(base.ComponentExecution{
+			Component: cmp,
+			Setup:     setup,
+			Task:      taskUpsert,
+		})
 		c.Assert(err, qt.IsNil)
 
 		pbIn := new(structpb.Struct)
-		_, err = exec.Execution.Execute(ctx, []*structpb.Struct{pbIn})
+		_, err = exec.Execute(ctx, []*structpb.Struct{pbIn})
 		c.Check(err, qt.IsNotNil)
 
 		want := "Pinecone responded with a 400 status code. Cannot provide both ID and vector at the same time."
@@ -271,11 +279,15 @@ func TestComponent_Execute(t *testing.T) {
 			"url": "http://no-such.host",
 		})
 
-		exec, err := connector.CreateExecution(nil, setup, taskUpsert)
+		exec, err := cmp.CreateExecution(base.ComponentExecution{
+			Component: cmp,
+			Setup:     setup,
+			Task:      taskUpsert,
+		})
 		c.Assert(err, qt.IsNil)
 
 		pbIn := new(structpb.Struct)
-		_, err = exec.Execution.Execute(ctx, []*structpb.Struct{pbIn})
+		_, err = exec.Execute(ctx, []*structpb.Struct{pbIn})
 		c.Check(err, qt.IsNotNil)
 
 		want := "Failed to call http://no-such.host/.*. Please check that the connector configuration is correct."